chore: move markmind module and add ask AI API
