+++ docs/deployments/systemd-agent.md
@@ -49,7 +49,14 @@ server {
     add_header Cache-Control "public";
   }
 
-  # 6. 隐藏 . 文件（如 .DS_Store）
+  # 6. 转发后端 API
+  location /api/ {
+    proxy_pass http://127.0.0.1:8080/api/;
+    proxy_set_header Host $host;
+    proxy_set_header X-Real-IP $remote_addr;
+  }
+
+  # 7. 隐藏 . 文件（如 .DS_Store）
   location ~ /\. {
     deny all;
   }
