+++ server/markmind/ingest/embed.go
